phase-2/backend/pyproject.toml
[project]
name = "todo-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "fastapi>=0.128.0",
    "pydantic-settings>=2.12.0",
    "python-jose>=3.5.0",
    "sqlmodel>=0.0.31",
    "uvicorn[standard]>=0.40.0",
]
